Optimize File System Management for Docker Launcher and Resolve class_registry Dependency Issue (#203)
* feat: Create TEMP_DIR if it doesn't exist

* chore: Update Docker Compose configuration and Dockerfile for MindSearch Docker launcher

* feat: Update Docker Compose configuration and Dockerfile modifying logic

* feat: Add new translations for current containers stop failure and container stopped and removed

* feat: Update Docker Compose configuration and Dockerfile for MindSearch Docker launcher

* Refactor Docker launcher's configuration setup and file handling

This commit refactors the configuration setup and file handling in the Docker launcher. It introduces a new `FileSystemManager` class with methods to ensure the existence of directories and files. The `ensure_dir` method is used to create the `temp` directory if it doesn't exist, and the `ensure_file` method is used to create the `.env` file with default content if it doesn't exist. This improves the reliability and maintainability of the Docker launcher.

* Add class_registry and python-dotenv lib

* Refactor Docker launcher's dependency installation and file handling

* Update docker/msdl/templates/backend/cloud_llm.dockerfile

Co-authored-by: liukuikun <24622904+Harold-lkk@users.noreply.github.com>

* cleanup dependencies & update MSDL version & enhance UX

---------

Co-authored-by: liukuikun <24622904+Harold-lkk@users.noreply.github.com>